+++ includes/content-gate/content-gifting/class-content-gifting.php
@@ -41,6 +41,7 @@ class Content_Gifting {
 		add_action( 'init', [ __CLASS__, 'hook_gift_button' ] );
 		add_action( 'wp', [ __CLASS__, 'unrestrict_content' ], 5 );
 		add_filter( 'newspack_content_gate_restrict_post', [ __CLASS__, 'restrict_post' ] );
+		add_filter( 'newspack_content_gate_metering_short_circuit', [ __CLASS__, 'short_circuit_metering' ] );
 		add_action( 'wp_enqueue_scripts', [ __CLASS__, 'enqueue_assets' ] );
 		add_action( 'admin_enqueue_scripts', [ __CLASS__, 'enqueue_assets' ] );
 		add_action( 'wp_ajax_' . self::GENERATE_ACTION, [ __CLASS__, 'ajax_generate_content_key' ] );
@@ -155,7 +156,7 @@ class Content_Gifting {
 	 * Whether the current post has been gifted.
 	 *
 	 * @param int|null    $post_id The post ID. Default is the current post.
-	 * @param string|null $key     The content key. Default is the key from the query arg.
+	 * @param string|null $key     The content key. Default is from the request (query arg or cookie).
 	 *
 	 * @return bool
 	 */
@@ -396,26 +397,6 @@ class Content_Gifting {
 			$errors->add( 'not_enabled', __( 'Content gifting is not enabled.', 'newspack-plugin' ) );
 		}
 
-		// Check whether all gates have metering enabled.
-		$gates = array_filter(
-			Content_Gate::get_gates(),
-			function( $gate ) {
-				return $gate['status'] === 'publish';
-			}
-		);
-		if ( ! empty( $gates ) ) {
-			$all_gates_have_metering = true;
-			foreach ( $gates as $gate ) {
-				if ( ! isset( $gate['metering']['enabled'] ) || ! $gate['metering']['enabled'] ) {
-					$all_gates_have_metering = false;
-					break;
-				}
-			}
-			if ( $all_gates_have_metering ) {
-				$errors->add( 'all_gates_have_metering', __( 'Content gifting is not available because all gates have metering enabled.', 'newspack-plugin' ) );
-			}
-		}
-
 		if ( $return_errors ) {
 			return $errors;
 		}
@@ -454,21 +435,6 @@ class Content_Gifting {
 		 * @param bool $enabled Whether the content gifting is enabled.
 		 */
 		do_action( 'newspack_content_gifting_enabled_status_changed', $enabled );
-	}
-
-	/**
-	 * Whether to render the metering notice in the configuration wizard.
-	 *
-	 * @return bool
-	 */
-	public static function should_render_metering_notice() {
-		$gates = Content_Gate::get_gates();
-		foreach ( $gates as $gate ) {
-			if ( $gate['status'] === 'publish' && isset( $gate['metering'] ) && $gate['metering']['enabled'] ) {
-				return true;
-			}
-		}
-		return false;
 	}
 
 	/**
@@ -674,6 +640,20 @@ class Content_Gifting {
 	}
 
 	/**
+	 * Short-circuit the metering check.
+	 *
+	 * @param mixed $short_circuit Short-circuit value. Default is null.
+	 *
+	 * @return mixed Short-circuit value.
+	 */
+	public static function short_circuit_metering( $short_circuit ) {
+		if ( self::is_gifted_post( get_the_ID() ) ) {
+			return true;
+		}
+		return $short_circuit;
+	}
+
+	/**
 	 * Whether the current user can gift a given post.
 	 *
 	 * @param int|null $post_id       Optional post ID. Default is the current post.
@@ -699,10 +679,6 @@ class Content_Gifting {
 
 		if ( Content_Gate::is_post_restricted( $post_id ) ) {
 			$errors->add( 'post_restricted', __( 'User does not have access to this post.', 'newspack-plugin' ) );
-		}
-
-		if ( Metering::has_metering( $post_id ) ) {
-			$errors->add( 'metering', __( 'Metered content cannot be gifted.', 'newspack-plugin' ) );
 		}
 
 		if ( $return_errors ) {