+++ src/wizards/audience/views/setup/content-gating.js
@@ -66,6 +66,7 @@ export default withWizardScreen( ( { wizardApiFetch } ) => {
 	};
 
 	const giftingErrors = Object.values( newspackAudience?.content_gifting?.can_use_gifting?.errors || {} ).flat();
+	const hasMetering = newspackAudience?.content_gifting?.has_metering;
 
 	return (
 		<WizardsTab
@@ -98,15 +99,6 @@ export default withWizardScreen( ( { wizardApiFetch } ) => {
 				{ config.content_gifting?.enabled && (
 					<>
 						{ giftingErrors.length > 0 && <Notice noticeText={ giftingErrors.join( ', ' ) } isError /> }
-						{ ! giftingErrors.length && newspackAudience.content_gifting.metering_notice && (
-							<Notice
-								noticeText={ __(
-									'You have a content gate with metering enabled. Mind that metered articles are not eligible for gifting.',
-									'newspack-plugin'
-								) }
-								isWarning
-							/>
-						) }
 						<Grid columns={ 2 } rowGap={ 32 }>
 							<Heading level={ 4 } style={ { gridColumn: '1 / -1' } }>
 								{ __( 'General Settings', 'newspack-plugin' ) }
@@ -207,7 +199,7 @@ export default withWizardScreen( ( { wizardApiFetch } ) => {
 							</ToggleGroupControl>
 							<div style={ { gridColumn: '1 / -1' } }>
 								<BaseControl id="newspack-content-gifting-cta-preview" label={ __( 'Preview', 'newspack-plugin' ) }>
-									<div className="newspack-content-gifting__cta-preview">
+									<div className="newspack-content-gifting__cta-preview" inert>
 										<div className="newspack-ui">
 											<div
 												className={ `banner newspack-content-gifting__cta is-style-${
@@ -215,13 +207,22 @@ export default withWizardScreen( ( { wizardApiFetch } ) => {
 												}` }
 											>
 												<div className="wrapper newspack-content-gifting__cta__content">
-													<span className="newspack-ui__font--s">
+													<div className="newspack-ui__font--s">
 														{ config.content_gifting.cta_label ||
 															__(
 																'This article has been gifted to you by someone who values great journalism.',
 																'newspack-plugin'
+															) }{ ' ' }
+														<div className="newspack-ui__font--xs newspack-content-gifting__cta__content__links">
+															{ hasMetering ? (
+																<a href="#register_modal">{ __( 'Create an account', 'newspack-plugin' ) }</a>
+															) : (
+																<a href="#signin_modal">
+																	{ __( 'Sign in to an existing account', 'newspack-plugin' ) }
+																</a>
 															) }
-													</span>
+														</div>
+													</div>
 													<button
 														className={ `newspack-ui__button newspack-ui__button--x-small ${
 															( config.content_gifting.style || 'light' ) === 'dark'